requirements.txt
streamlit>=1.42,<2.0
gspread>=6.0,<7.0
google-api-python-client>=2.120,<3.0
google-auth>=2.28,<3.0
google-auth-oauthlib>=1.2,<2.0
Authlib>=1.3,<2.0
pandas>=2.2,<3.0